+++ scripts/xmake/verilator.lua
@@ -39,42 +39,44 @@ function emu_comp(num_cores)
   local difftest_csrc_verilator = path.join(difftest_csrc, "verilator")
   local difftest_config = path.join(difftest, "config")
 
-  depend.on_changed(function ()
-    if os.exists(build_dir) then os.rmdir(build_dir) end
-    task.run("soc", {
-      sim = true, config = option.get("config"),
-      dramsim3 = option.get("dramsim3"), enable_perf = not option.get("no_perf"),
-      socket = option.get("socket"), lua_scoreboard = option.get("lua_scoreboard"),
-      core = option.get("core"), l3 = option.get("l3"), noc = option.get("noc"),
-      legacy = option.get("legacy"), jar = option.get("jar"),
-      build_dir = build_dir
-    })
-    local vsrc = os.files(path.join(design_vsrc, "*v"))
-    table.join2(vsrc, os.files(path.join(difftest_vsrc_common, "*v")))
-    table.join2(vsrc, os.files(path.join(difftest_vsrc_st, "*v")))
+  if not option.get("no_build_chisel") then
+    depend.on_changed(function ()
+      if os.exists(build_dir) then os.rmdir(build_dir) end
+      task.run("soc", {
+        sim = true, config = option.get("config"),
+        dramsim3 = option.get("dramsim3"), enable_perf = not option.get("no_perf"),
+        socket = option.get("socket"), lua_scoreboard = option.get("lua_scoreboard"),
+        core = option.get("core"), l3 = option.get("l3"), noc = option.get("noc"),
+        legacy = option.get("legacy"), jar = option.get("jar"),
+        build_dir = build_dir
+      })
+      local vsrc = os.files(path.join(design_vsrc, "*v"))
+      table.join2(vsrc, os.files(path.join(difftest_vsrc_common, "*v")))
+      table.join2(vsrc, os.files(path.join(difftest_vsrc_st, "*v")))
 
-    if option.get("lua_scoreboard") then
-      local dpi_cfg_lua = path.join(abs_base, "scripts", "verilua", "dpi_cfg.lua")
-      if os.exists(dpi_export_dir) then os.rmdir(dpi_export_dir) end
-      os.mkdir(dpi_export_dir)
-      local dpi_exp_opts =  {"dpi_exporter"}
-      table.join2(dpi_exp_opts, {"--config", dpi_cfg_lua})
-      table.join2(dpi_exp_opts, {"--out-dir", dpi_export_dir})
-      table.join2(dpi_exp_opts, {"--work-dir", dpi_export_dir})
-      table.join2(dpi_exp_opts, {"-I", design_gen_dir})
-      table.join2(dpi_exp_opts, {"--quiet"})
-      table.join2(dpi_exp_opts, {"--top", vtop})
-      table.join2(dpi_exp_opts, vsrc)
-      local cmd_file = path.join(comp_dir, "dpi_exp_cmd.sh")
-      io.writefile(cmd_file, table.concat(dpi_exp_opts, " "))
-      os.execv(os.shell(), { cmd_file })
-    end
-  end,{
-    files = chisel_dep_srcs,
-    dependfile = path.join("out", "chisel.verilator.dep." .. (build_dir):gsub("/", "_"):gsub(" ", "_")),
-    dryrun = option.get("rebuild"),
-    values = table.join2({build_dir}, xmake.argv())
-  })
+      if option.get("lua_scoreboard") then
+        local dpi_cfg_lua = path.join(abs_base, "scripts", "verilua", "dpi_cfg.lua")
+        if os.exists(dpi_export_dir) then os.rmdir(dpi_export_dir) end
+        os.mkdir(dpi_export_dir)
+        local dpi_exp_opts =  {"dpi_exporter"}
+        table.join2(dpi_exp_opts, {"--config", dpi_cfg_lua})
+        table.join2(dpi_exp_opts, {"--out-dir", dpi_export_dir})
+        table.join2(dpi_exp_opts, {"--work-dir", dpi_export_dir})
+        table.join2(dpi_exp_opts, {"-I", design_gen_dir})
+        table.join2(dpi_exp_opts, {"--quiet"})
+        table.join2(dpi_exp_opts, {"--top", vtop})
+        table.join2(dpi_exp_opts, vsrc)
+        local cmd_file = path.join(comp_dir, "dpi_exp_cmd.sh")
+        io.writefile(cmd_file, table.concat(dpi_exp_opts, " "))
+        os.execv(os.shell(), { cmd_file })
+      end
+    end,{
+      files = chisel_dep_srcs,
+      dependfile = path.join("out", "chisel.verilator.dep." .. (build_dir):gsub("/", "_"):gsub(" ", "_")),
+      dryrun = option.get("rebuild"),
+      values = table.join2({build_dir}, xmake.argv())
+    })
+  end
 
   assert(#os.files(path.join(design_vsrc, "*v")) > 0, "[verilator.lua] [emu_comp] rtl dir(`%s`) is empty!", design_vsrc)
